refactor(trackerdash): Massage code until TrackerDash example runs as Candela example
