+++ app/examples/index.json
@@ -107,5 +107,9 @@
   {
     "title": "OnSet",
     "link": "onset"
+  },
+  {
+    "title": "TrackerDash",
+    "link": "trackerdash"
   }
 ]
