+++ app/examples/trackerdash/index.jade
@@ -6,5 +6,4 @@ head
 body
   .dash-container
 
-  script(src="dist/trackerdash.widget.js" charset="utf-8")
   script(src="index.js")
