+++ package.json
@@ -60,6 +60,8 @@
     "babel-preset-es2015": "^6.3.13",
     "backbone": "^1.2.3",
     "bluebird": "^3.3.5",
+    "bootstrap": "^3.3.7",
+    "bootstrap-webpack": "0.0.5",
     "callsite": "^1.0.0",
     "clean-webpack-plugin": "^0.1.7",
     "codecov.io": "^0.1.6",
@@ -74,6 +76,7 @@
     "eslint-plugin-standard": "^1.3.2",
     "exports-loader": "^0.6.3",
     "expose-loader": "^0.7.1",
+    "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
     "geojs": "^0.9.1",
     "grunt": "^1.0.1",
@@ -81,6 +84,7 @@
     "html-webpack-plugin": "^2.9.0",
     "http-server": "^0.9.0",
     "husky": "^0.11.8",
+    "imports-loader": "^0.6.5",
     "jade": "^1.11.0",
     "jade-loader": "^0.8.0",
     "javascript-detect-element-resize": "^0.5.3",
@@ -94,9 +98,11 @@
     "karma-webpack": "^1.7.0",
     "legacy-loader": "0.0.2",
     "less": "^2.3.1",
+    "md5": "^2.2.1",
     "nightmare": "^2.3.1",
     "node-resemble": "^1.1.3",
     "node-sass": "^3.4.2",
+    "nvd3": "^1.8.4",
     "pegjs": "^0.9.0",
     "pegjs-loader": "^0.4.0",
     "phantomjs-prebuilt": "^2.1.5",
