+++ src/candela/components/TrackerDash/templates/infoPane.jade
