+++ src/candela/components/TrackerDash/templates/statusBarWidget.jade
