+++ src/candela/components/TrackerDash/templates/tablePane.jade
