+++ src/candela/components/TrackerDash/templates/trendPane.jade
